added fixing types sizes/byte_offsets to uint32_t and element_indexes and tuple_indexes to uint16_t

+++ inc/tuple.h
@@ -8,27 +8,27 @@
 // prior to calling any of the following functions below
 
 // element level functions
-uint64_t get_element_size(const tuple_def* tpl_d, uint64_t index, const void* tupl);
+uint32_t get_element_size(const tuple_def* tpl_d, uint16_t index, const void* tupl);
 
-uint64_t get_element_offset(const tuple_def* tpl_d, uint64_t index, const void* tupl);
+uint32_t get_element_offset(const tuple_def* tpl_d, uint16_t index, const void* tupl);
 
-element seek_to_element(const tuple_def* tpl_d, uint64_t index, const void* tupl);
+element seek_to_element(const tuple_def* tpl_d, uint16_t index, const void* tupl);
 
 
 // tuple level functions
-uint64_t get_tuple_size(const tuple_def* tpl_d, const void* tupl);
+uint32_t get_tuple_size(const tuple_def* tpl_d, const void* tupl);
 
 void* seek_to_end_of_tuple(const tuple_def* tpl_d, const void* tupl);
 
 
 // copy element to and from tuple functions
-void copy_element_to_tuple(const tuple_def* tpl_d, uint64_t index, void* tupl, const void* value);
+void copy_element_to_tuple(const tuple_def* tpl_d, uint16_t index, void* tupl, const void* value);
 
-void copy_element_from_tuple(const tuple_def* tpl_d, uint64_t index, const void* tupl, void* value);
+void copy_element_from_tuple(const tuple_def* tpl_d, uint16_t index, const void* tupl, void* value);
 
 
 // compare functions for tuple and for an element inside the tuple
-int compare_elements(void* tup1, void* tup2, const tuple_def* tpl_d, uint64_t index);
+int compare_elements(void* tup1, void* tup2, const tuple_def* tpl_d, uint16_t index);
 
 int compare_tuples(void* tup1, void* tup2, const tuple_def* tpl_d);
